requirements.txt
cycler==0.11.0
fonttools==4.33.3
imageio==2.19.3
kiwisolver==1.4.3
matplotlib==3.5.2
networkx==2.8.4
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
tifffile==2022.5.4
